API controller logic move to view, minor bug fixed (merchantController and merchant model)
